+++ CHANGELOG.md
@@ -13,6 +13,7 @@ ChangeLog
 - IMPROVEMENT: add "compressor" audio processing node
 - IMPROVEMENT: add "awstranslate" text translation node
 - IMPROVEMENT: add "awspolly" test to speech generation node
+- IMPROVEMENT: add "openaitranscribe" speech to text node
 
 1.4.5 (2025-08-07)
 ------------------